+++ fusion/src/Fusion/autodiff/policies/Operation.h
@@ -14,16 +14,16 @@ template <typename T>
 struct UnaryType {std::vector<T> a;};
 
 template <typename T, typename Enable = void>
-struct static_arity : std::integral_constant<std::uint16_t, 0> {};
+struct static_arity : std::integral_constant<std::size_t, 0> {};
 
 
 template <typename T>
 struct static_arity<UnaryType<T>, void>
-    : std::integral_constant<std::uint16_t, 1> {};
+    : std::integral_constant<std::size_t, 1> {};
 
 template <typename T>
 struct static_arity<BinaryType<T>, void>
-    : std::integral_constant<std::uint16_t, 2> {};
+    : std::integral_constant<std::size_t, 2> {};
 
 template <typename T>
 struct Context {
